Refactor remaining selectors
